requirements.txt
streamlit
pandas
numpy
plotly.express
os
datetime